Correct deleted_at value in notification messages
The deleted_at value in notification messages is currently
set from the terminated_at value in the instance record.

A comment in the code says this is to avoid confusion between
terminated_at and deleted_at,  but swapping the names doesn't
seem to be a great way to address that.

In practice "terminated and not deleted" is a transient state
unless an error occurs, and the time gap is very short.  But
deleted_at is always set when the entry is deleted in the DB,
whereas terminated_at can be left unset in some error cases.

Rather than present terminated_at as the deleted_at value in
the message we include both and let the recipient decide which
they want to use.

Fixes bug: 1227080

Change-Id: Ic67f66b6c64be1a8f43f4fdd5b1fac25f1414247