Get gantt to pass unit tests
Mostly a matter of having the appropriate modules imported from
the gantt tree rather than the nova tree.

Note that code can still import from the nova tree but this is only
for non-scheduler related files.  Ultimately we will remove all
nova references from the gantt code but that is a future project.

With these changes we can now have Jenkins gate on the unit tests.

Change-Id: I3bcc9742cc9ab4188f67c2aa66ad97eca2bc08bb